+++ src/main/java/absolutelyaya/ultracraft/entity/projectile/CancerBulletEntity.java
@@ -5,10 +5,13 @@ import absolutelyaya.ultracraft.accessor.ProjectileEntityAccessor;
 import absolutelyaya.ultracraft.damage.DamageSources;
 import absolutelyaya.ultracraft.registry.EntityRegistry;
 import absolutelyaya.ultracraft.registry.ItemRegistry;
+import absolutelyaya.ultracraft.registry.StatusEffectRegistry;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.projectile.thrown.ThrownItemEntity;
 import net.minecraft.item.Item;
+import net.minecraft.util.hit.EntityHitResult;
 import net.minecraft.util.hit.HitResult;
 import net.minecraft.world.World;
 
@@ -19,6 +22,7 @@ public class CancerBulletEntity extends HellBulletEntity implements ProjectileEn
 	public CancerBulletEntity(EntityType<? extends ThrownItemEntity> entityType, World world)
 	{
 		super(entityType, world);
+		damageSource = DamageSources.get(world, DamageSources.CANCER, this, this.getOwner());
 	}
 	
 	protected CancerBulletEntity(LivingEntity owner, World world)