+++ src/main/java/absolutelyaya/ultracraft/mixin/client/render/LivingEntityRendererMixin.java
@@ -1,17 +1,29 @@
 package absolutelyaya.ultracraft.mixin.client.render;
 
-import absolutelyaya.ultracraft.Ultracraft;
-import net.minecraft.client.render.WorldRenderer;
+import absolutelyaya.ultracraft.registry.StatusEffectRegistry;
+import net.minecraft.client.render.VertexConsumerProvider;
+import net.minecraft.client.render.entity.LivingEntityRenderer;
+import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.entity.LivingEntity;
+import org.spongepowered.asm.mixin.Debug;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.ModifyArg;
+import org.spongepowered.asm.mixin.injection.ModifyArgs;
+import org.spongepowered.asm.mixin.injection.invoke.arg.Args;
 
-@Mixin(WorldRenderer.class)
-public class LivingEntityRendererMixin
+@Mixin(LivingEntityRenderer.class)
+public class LivingEntityRendererMixin<T extends LivingEntity>
 {
-	@ModifyArg(method = "render", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/WorldRenderer;renderEntity(Lnet/minecraft/entity/Entity;DDDFLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;)V"), index = 4)
-	float adjustTickDelta(float tickDelta)
+	@ModifyArgs(method = "render(Lnet/minecraft/entity/LivingEntity;FFLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;I)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/entity/model/EntityModel;render(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumer;IIFFFF)V"))
+	void modifyRenderColor(Args args, T livingEntity, float f, float g, MatrixStack matrixStack, VertexConsumerProvider vertexConsumerProvider, int i)
 	{
-		return Ultracraft.isTimeFrozen() ? 0f : tickDelta;
+		//TODO: other entities effects don't get networked; meaning this only works on client player entities...
+		if(livingEntity.hasStatusEffect(StatusEffectRegistry.CANCEROUS))
+		{
+			args.set(4, 0.2f);
+			args.set(5, 1f);
+			args.set(6, 0.3f);
+			args.set(7, 0.5f);
+		}
 	}
 }