+++ src/main/java/absolutelyaya/ultracraft/mixin/client/render/WorldRendererMixin.java
@@ -1,10 +1,15 @@
 package absolutelyaya.ultracraft.mixin.client.render;
 
+import absolutelyaya.ultracraft.Ultracraft;
 import absolutelyaya.ultracraft.block.SkyBlockEntity;
 import absolutelyaya.ultracraft.client.RenderLayers;
 import absolutelyaya.ultracraft.client.UltracraftClient;
+import absolutelyaya.ultracraft.registry.StatusEffectRegistry;
+import com.mojang.blaze3d.systems.RenderSystem;
 import net.minecraft.client.render.*;
 import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.util.math.Vec3d;
 import org.joml.Matrix4f;
 import org.spongepowered.asm.mixin.Debug;
@@ -13,9 +18,9 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Debug(export = true)
 @Mixin(WorldRenderer.class)
 public abstract class WorldRendererMixin
 {
@@ -38,4 +43,17 @@ public abstract class WorldRendererMixin
 		for (SkyBlockEntity.SkyType type : SkyBlockEntity.SkyType.values())
 			bufferBuilders.getEntityVertexConsumers().draw(RenderLayers.getSky(type));
 	}
+	
+	@ModifyArg(method = "render", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/render/WorldRenderer;renderEntity(Lnet/minecraft/entity/Entity;DDDFLnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;)V"), index = 4)
+	float adjustTickDelta(float tickDelta)
+	{
+		return Ultracraft.isTimeFrozen() ? 0f : tickDelta;
+	}
+	
+	//@Inject(method = "renderEntity", at = @At("HEAD"))
+	//void beforeRender(Entity entity, double cameraX, double cameraY, double cameraZ, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, CallbackInfo ci)
+	//{
+	//	if(entity instanceof LivingEntity living && living.hasStatusEffect(StatusEffectRegistry.CANCEROUS))
+	//		RenderSystem.setShaderColor(0.2f, 1f, 0.3f, 1f);
+	//}
 }